Add files via upload
bag_parser.m processes a .bag file from ROS, and converts the structured array of messages from one topic into an ordered numerical array of messages from any type. Note that a valid .bag file must be included in the environment folder, and the filename must be appropriately specified in the first line of the script. A 3x1 subplot is produced in this version, but the script can be edited for whatever computation is needed.